Finish reorganizing common schema code
Also modified fallback schema test to load fallback schema to determine
its version, avoiding a second code change when the schema is updated.
The falsy Memoizer key issue (a comment from #51) is fixed here too.
