+++ common/schema/config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  /** Path to the fallback HED schema. */
+  fallbackFilePath: 'data/HED8.0.0.xml',
+}
